remove typecating

+++ bit_manipulation/1-print_binary.c
@@ -17,7 +17,7 @@ void print_binary(unsigned long int n)
 	if (n == 0)
 		_putchar('0');
 	leading_zero = 1; /* keep track of leading zeroes in n */
-	bitmask = ((unsigned long int)1) << 31; /* for 64-bit integers */
+	bitmask = 1 << 31; /* for 32-bit integers */
 	for (i = 0; i < 32; i++)
 	{
 		/* use '&' to find the first non-zero binary in n */
